backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
google-genai==0.3.0
openai==1.14.0
langchain==0.1.13
langchain-google-genai==0.0.11
langchain-openai==0.0.8
pytest==8.1.1
httpx==0.27.0
